requirements.txt
colorama>=0.4.6
requests>=2.28.0
pdfplumber>=0.9.0
python-dotenv>=1.0.0
